renaming 'definitions' folder to 'models'
