+++ NodeJS/src/index.ts
@@ -1,3 +1,3 @@
 export * from './AgentClient';
 export * from './SubscriptionClient';
-export * from './definitions';
+export * from './models';
